new(cmd): set wildcard as default name for directories and files.
Moreover, add unit tests for shallow finds.
